# swagger--old.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Book Api Do Jpzin",
    "description": "API for managing books",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/register":{
      "post":{
        "summary":"Create a User",
        "tags":[
          "Auth"
        ],
        "requestBody":{
          "required":true,
          "content":{
            "application/json":{
              "schema":{
                "type":"object",
                "properties":{
                  "username":{
                    "type":"string"
                  },
                  "password":{
                    "type":"string"
                  }
                }
              }
            }
          }
        },
        "responses":{
          "201":{
            "description":"User created sucessfully",
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "message":{
                      "type":"string"
                    }
                  }
                }
              }

            }
          },
          "400":{
            "description":"Bad Request"
          }
        }
        
      }
    },
    "/auth/login":{
      "post":{
        "summary":"Login a User",
        "tags":[
          "Auth"
        ],
        "requestBody":{
          "required":true,
          "content":{
            "application/json":{
              "schema":{
                "type":"object",
                "properties":{
                  "username":{
                    "type":"string"
                  },
                  "password":{
                    "type":"string"
                  }
                }
              }
            }
          }
        },
        "responses":{
          "200":{
            "description":"Login sucessfully",
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "token":{
                      "type":"string"
                    }
                  }
                }
              }
            }
          },
          "401":{
            "description":"Invalid credentials"
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}